+++ lib/fast_jsonapi/scalar.rb
@@ -10,17 +10,17 @@ module FastJsonapi
 
     def serialize(record, serialization_params, output_hash)
       if conditionally_allowed?(record, serialization_params)
-        output_hash[key] = if method.is_a?(Proc)
-          method.arity.abs == 1 ? method.call(record) : method.call(record, serialization_params)
+        if method.is_a?(Proc)
+          output_hash[key] = FastJsonapi.call_proc(method, record, serialization_params)
         else
-          record.public_send(method)
+          output_hash[key] = record.public_send(method)
         end
       end
     end
 
     def conditionally_allowed?(record, serialization_params)
       if conditional_proc.present?
-        conditional_proc.call(record, serialization_params)
+        FastJsonapi.call_proc(conditional_proc, record, serialization_params)
       else
         true
       end